requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
djangorestframework==3.15.2
mysqlclient==2.2.4
python-dotenv==1.0.1
tzdata==2024.1